+++ egui/src/containers/area.rs
@@ -172,10 +172,11 @@ impl Prepared {
 
         state.size = content_ui.min_rect().size();
 
-        let interact_id = if movable {
-            Some(layer_id.id.with("move"))
+        let interact_id = layer_id.id.with("move");
+        let sense = if movable {
+            Sense::click_and_drag()
         } else {
-            None
+            Sense::click() // allow clicks to bring to front
         };
 
         let move_response = ctx.interact(
@@ -184,16 +185,16 @@ impl Prepared {
             ctx.style().spacing.item_spacing,
             state.rect(),
             interact_id,
-            Sense::click_and_drag(),
+            sense,
         );
 
-        if move_response.active {
+        if move_response.active && movable {
             state.pos += ctx.input().mouse.delta;
         }
 
         state.pos = ctx.constrain_window_rect(state.rect()).min;
 
-        if move_response.active
+        if (move_response.active || move_response.clicked)
             || mouse_pressed_on_area(ctx, layer_id)
             || !ctx.memory().areas.visible_last_frame(&layer_id)
         {